# docs/openapi/admin.yaml
openapi: 3.0.3
...
info:
  title: voxeltoad — Admin API
  version: 0.1.0
  description: >
    Management-plane REST API for providers, models, routes, plugins, tenants,
    API keys, quotas, usage, and audit. Consumed by the Control Panel UI and the
    generated TypeScript admin SDK.
servers:
  - url: /
    description: Same-origin (admin plane root)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/request-logs:
    get:
      tags: [request-logs]
      summary: List data-plane request-audit entries (offset paginated, time-bounded)
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: tenant
          in: query
          required: false
          schema: { type: string }
          description: super-admin only; filter to a single tenant. Ignored for tenant-admin (implicitly own tenant).
        - name: group_name
          in: query
          required: false
          schema: { type: string }
        - name: api_key_id
          in: query
          required: false
          schema: { type: string }
        - name: provider
          in: query
          required: false
          schema: { type: string }
        - name: model_requested
          in: query
          required: false
          schema: { type: string }
        - name: error_type
          in: query
          required: false
          schema: { type: string }
        - name: blocked_by
          in: query
          required: false
          schema: { type: string }
        - name: stream
          in: query
          required: false
          schema: { type: boolean }
          description: Filter streaming vs non-streaming requests.
        - name: fallback
          in: query
          required: false
          schema: { type: boolean }
          description: Filter requests that failed over across candidates.
        - name: session_id
          in: query
          required: false
          schema: { type: string }
          description: Filter to requests belonging to a single session (X-Voxeltoad-Session header value).
        - name: request_id
          in: query
          required: false
          schema: { type: string }
          description: Filter to a single request by its gateway-assigned correlation id (exact lookup, debug/trace drill-down).
        - name: format
          in: query
          required: false
          schema: { type: string, enum: [csv] }
          description: Set to "csv" to export as CSV instead of JSON (2000-row cap, mirrors usage/audit CSV endpoints per ADR-0028).
      responses:
        '200':
          description: Request-log entry page (JSON) or CSV export (when format=csv)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RequestLogPage' }
            text/csv:
              schema: { type: string, description: "CSV export with headers: id,tenant,group_name,api_key_id,provider,model_requested,model_resolved,stream,prompt_tokens,completion_tokens,total_tokens,ttft_ms,duration_ms,error_type,blocked_by,fallback,request_id,session_id,trace_id,session_source,agent_type,cache_hit,cache_tier,cache_source,cached_prompt_tokens,created_at" }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/request-logs/sessions:
    get:
      tags: [request-logs]
      summary: Session list — aggregated per-session totals
      description: >
        Aggregates request_logs by session_id, returning per-session totals
        (request count, token sums, total duration, latest agent type,
        has-errors flag) plus cost merged from usage_records. Ordered by
        most-recent activity DESC. Drives the session-list view of the trace UI.
        A time window is advised to bound the GROUP BY scan. Tenant scope applies
        identically to /request-logs.
      parameters:
        - name: agent_type
          in: query
          required: false
          schema: { type: string }
          description: Filter sessions by detected agent (claude-code/codex/codebuddy/workbuddy/opencode).
        - name: from
          in: query
          required: false
          schema: { type: string, format: date-time }
          description: Inclusive lower bound on created_at (RFC3339).
        - name: to
          in: query
          required: false
          schema: { type: string, format: date-time }
          description: Exclusive upper bound on created_at (RFC3339).
        - name: page
          in: query
          required: false
          schema: { type: integer, default: 1, minimum: 1 }
        - name: page_size
          in: query
          required: false
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: Paginated list of session summaries.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/SessionSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/request-logs/sessions/{session_id}:
    get:
      tags: [request-logs]
      summary: Session trace — chronological request timeline + cost summary for a session
      description: >
        Returns all request_logs entries for a given session_id in chronological
        (ASC) order, plus the session's aggregated cost/tokens from
        usage_records. Enables the session-chain analysis view (debug + cost
        attribution). Tenant scope applies identically to /request-logs.
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
          description: The X-Voxeltoad-Session header value to trace.
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 200, maximum: 500 }
          description: Caps the number of request rows returned (bounds cost on long-lived sessions).
      responses:
        '200':
          description: Session trace (requests in timeline order + cost summary).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionTrace' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ---------------------------------------------------------------------------
  # LLM trace payloads — the message + raw layers below request_logs (ADR-0039).
  # Where request_logs stores only metadata, trace_payloads stores the
  # prompt/completion plaintext and raw upstream request/response bodies,
  # enabling the Session → Request → Messages → Raw debugging view. Capture is
  # opt-in on the data plane; these reads are RBAC + tenant-scoped, and the
  # detail read (plaintext) is audited.
  # ---------------------------------------------------------------------------
...
  /api/v1/permissions:
    get:
      tags: [roles]
      summary: List all available permission points (catalog)
      responses:
        '200':
          description: Permission catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Permission'
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/roles:
    get:
      tags: [roles]
      summary: List all roles with their permission sets
      responses:
        '200':
          description: Role list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RoleWithPermissions'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [roles]
      summary: Create a new custom role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Created role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleWithPermissions'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/roles/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    patch:
      tags: [roles]
      summary: Update a role's description and/or permissions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Updated role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleWithPermissions'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [roles]
      summary: Delete a custom role (built-in roles cannot be deleted)
      responses:
        '204':
          description: Role deleted
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque operator session token from POST /auth/login.
  parameters:
...
    From:
      name: from
      in: query
      required: false
      schema: { type: string, format: date-time }
      description: Inclusive lower bound on created_at (RFC3339).
    To:
      name: to
      in: query
      required: false
      schema: { type: string, format: date-time }
      description: Exclusive upper bound on created_at (RFC3339).
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 500, default: 1 }
      description: 1-based page number for offset pagination (clamped to max 500 to bound deep-offset scans).
    PageSize:
      name: page_size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Number of items per page.
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing/invalid/expired session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not authorized for this resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found (or not in caller's tenant)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Resource conflict (e.g. role still referenced by operators)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message: { type: string }
            type:
              type: string
              description: OpenAI-compatible error type.
              enum:
                - invalid_request_error
                - authentication_error
                - permission_error
                - not_found
                - rate_limit_error
                - api_error
...
    PageEnvelope:
      type: object
      description: Offset-pagination envelope for the audit & request-logs page-jump endpoints (separate from the keyset ListEnvelope to avoid perturbing ADR-0019).
      required: [data, total, page, page_size]
      properties:
        data:
          type: array
          items: {}
        total:
          type: integer
          format: int64
          description: Total matching rows for the current filter (independent of page).
        page:
          type: integer
          description: Current 1-based page number.
        page_size:
          type: integer
          description: Page size used for this response.

    # --- Money ---
...
    Permission:
      type: object
      description: One permission entry from the catalog (authz.AllPermissions).
      properties:
        perm: { type: string, description: "Permission key (e.g. provider.write)." }
        scope: { type: string, enum: [global, tenant], description: "Whether this permission is global-scope or tenant-scope." }
        label: { type: string, description: "Human-readable label." }
    RoleWithPermissions:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        scope_kind: { type: string, enum: [global, tenant] }
        is_builtin: { type: boolean }
        description: { type: string }
        permissions:
          type: array
          items: { type: string }
    CreateRoleRequest:
      type: object
      required: [name, scope_kind, permissions]
      properties:
        name: { type: string }
        scope_kind: { type: string, enum: [global, tenant] }
        description: { type: string }
        permissions:
          type: array
          items: { type: string }
          description: "Permission keys for this role."
    UpdateRoleRequest:
      type: object
      description: >
        At least one of description or permissions must be provided.
        Built-in roles cannot change name/scope_kind.
      properties:
        description: { type: string }
        permissions:
          type: array
          items: { type: string }

    # --- Tenancy ---
...
    RequestLogEntry:
      type: object
      description: >
        One data-plane request record: identity, provider/model routing
        outcome, token/timing facts, and error/blocking classification. Not to
        be confused with AuditEntry (management-plane mutation audit).
      properties:
        id: { type: integer, format: int64 }
        tenant: { type: string }
        group_name: { type: string }
        api_key_id: { type: string }
        provider: { type: string }
        model_requested: { type: string }
        model_resolved: { type: string, description: Upstream model name actually resolved to (ADR-0002). }
        stream: { type: boolean }
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        total_tokens: { type: integer }
        ttft_ms: { type: integer, description: "Time-to-first-token, milliseconds (0 for non-streaming)." }
        duration_ms: { type: integer }
        error_type: { type: string, description: "OpenAI-compatible error type, empty on success." }
        blocked_by: { type: string, description: Plugin name that blocked the request (empty if not blocked). }
        fallback: { type: boolean, description: Whether the request failed over across candidates before succeeding. }
        request_id: { type: string, description: "Gateway-assigned per-request correlation id (sourced from the X-Request-Id/X-Trace-Id/traceparent header when present, otherwise chi-generated)." }
        upstream_request_id: { type: string, description: "Provider-assigned request id from the upstream response (OpenAI x-request-id header, Anthropic request-id header/body, …). Final/successful attempt only; empty when no upstream call succeeded or the provider returned no id." }
        session_id: { type: string, description: "Client-supplied session key from X-Voxeltoad-Session header." }
        trace_id: { type: string, description: "W3C trace id parsed from the traceparent header (empty if absent/invalid)." }
        session_source: { type: string, description: "Which mechanism supplied the session key (header-config, header-generic, body-session, body-metadata, body-user, prefix, or empty)." }
        agent_type: { type: string, description: "Detected calling agent (claude-code/codex/codebuddy/workbuddy/opencode); empty when unrecognized." }
        created_at: { type: string, format: date-time }

    # --- Typed list envelopes (data item types pinned) ---
...
    RequestLogPage:
      allOf:
        - $ref: '#/components/schemas/PageEnvelope'
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/RequestLogEntry' } }
    SessionCostSummary:
      type: object
      description: Aggregated cost/tokens for all usage_records rows in a session.
      properties:
        session_id: { type: string }
        prompt_tokens: { type: integer, format: int64 }
        completion_tokens: { type: integer, format: int64 }
        cost: { type: integer, format: int64, description: int64 micro-units (ADR-0013). }
        request_count: { type: integer, format: int64 }
    SessionTrace:
      type: object
      description: >
        Session-chain analysis payload: the chronological request timeline (from
        request_logs) and the session's aggregated cost (from usage_records).
      properties:
        session_id: { type: string }
        requests:
          type: array
          items: { $ref: '#/components/schemas/RequestLogEntry' }
          description: Request-log entries in created_at ASC order (timeline order).
        cost_summary: { $ref: '#/components/schemas/SessionCostSummary' }
    SessionSummary:
      type: object
      description: >
        One row of the session-list aggregation: per-session totals over
        request_logs (tokens, duration, request count) plus the latest detected
        agent type and cost merged from usage_records. Drives the trace UI's
        session-list view.
      properties:
        session_id: { type: string }
        agent_type: { type: string, description: "Detected agent (claude-code/codex/…); empty when unrecognized." }
        request_count: { type: integer, description: Number of requests in the session. }
        prompt_tokens: { type: integer, description: Sum of prompt tokens across the session. }
        completion_tokens: { type: integer, description: Sum of completion tokens across the session. }
        total_tokens: { type: integer, description: Sum of total tokens across the session. }
        duration_ms: { type: integer, description: Sum of per-request durations in ms. }
        cost: { type: integer, format: int64, description: Total cost in micro-units (from usage_records). }
        started_at: { type: string, format: date-time, description: First request timestamp. }
        last_seen: { type: string, format: date-time, description: Most recent request timestamp. }
        has_errors: { type: boolean, description: Whether any request in the session errored. }

    # --- LLM trace payloads (ADR-0039): message + raw layers below request_logs.
    # trace_payloads stores prompt/completion plaintext + raw bodies; the summary
    # row omits the JSONB bodies so the request-list view renders cheaply, and
    # the detail row carries them for the Messages/Raw drill-down. ---
...